+++ library/src/actions/maxGraphemes/maxGraphemes.test-d.ts
@@ -30,19 +30,20 @@ describe('maxGraphemes', () => {
   });
 
   describe('should infer correct types', () => {
-    type Action = MaxGraphemesAction<string, 10, undefined>;
+    type Input = 'example string';
+    type Action = MaxGraphemesAction<Input, 10, undefined>;
 
     test('of input', () => {
-      expectTypeOf<InferInput<Action>>().toEqualTypeOf<string>();
+      expectTypeOf<InferInput<Action>>().toEqualTypeOf<Input>();
     });
 
     test('of output', () => {
-      expectTypeOf<InferOutput<Action>>().toEqualTypeOf<string>();
+      expectTypeOf<InferOutput<Action>>().toEqualTypeOf<Input>();
     });
 
     test('of issue', () => {
       expectTypeOf<InferIssue<Action>>().toEqualTypeOf<
-        MaxGraphemesIssue<string, 10>
+        MaxGraphemesIssue<Input, 10>
       >();
     });
   });