+++ library/src/actions/maxGraphemes/maxGraphemes.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, test } from 'vitest';
-import { _getGraphemes } from '../../utils/index.ts';
+import type { StringIssue } from '../../schemas/index.ts';
+import { _getGraphemeCount } from '../../utils/index.ts';
 import { expectActionIssue, expectNoActionIssue } from '../../vitest/index.ts';
 import {
   maxGraphemes,
@@ -9,17 +10,14 @@ import {
 
 describe('maxGraphemes', () => {
   describe('should return action object', () => {
-    const baseAction: Omit<
-      MaxGraphemesAction<string, 5, never>,
-      'message'
-    > = {
+    const baseAction: Omit<MaxGraphemesAction<string, 5, never>, 'message'> = {
       kind: 'validation',
       type: 'max_graphemes',
       reference: maxGraphemes,
       expects: '<=5',
       requirement: 5,
       async: false,
-      _run: expect.any(Function),
+      '~validate': expect.any(Function),
     };
 
     test('with undefined message', () => {
@@ -51,14 +49,31 @@ describe('maxGraphemes', () => {
     const action = maxGraphemes(5);
 
     test('for untyped inputs', () => {
-      expect(action._run({ typed: false, value: null }, {})).toStrictEqual({
+      const issues: [StringIssue] = [
+        {
+          kind: 'schema',
+          type: 'string',
+          input: null,
+          expected: 'string',
+          received: 'null',
+          message: 'message',
+        },
+      ];
+      expect(
+        action['~validate']({ typed: false, value: null, issues }, {})
+      ).toStrictEqual({
         typed: false,
         value: null,
+        issues,
       });
     });
 
     test('for valid strings', () => {
-      expectNoActionIssue(action, ['🧑🏻‍💻', '😀😀😀😀', 'foo']);
+      expectNoActionIssue(action, ['', ' ', '1', 'foo', '12345', '12 45']);
+    });
+
+    test('for valid emoji', () => {
+      expectNoActionIssue(action, ['😀', '😀👋🏼', '😀👋🏼🧩👩🏻‍🏫', '😀👋🏼🧩👩🏻‍🏫🫥']);
     });
   });
 
@@ -79,8 +94,17 @@ describe('maxGraphemes', () => {
       expectActionIssue(
         action,
         baseIssue,
-        ['😀😀😀😀😀😀', 'foobarbaz123'],
-        (value) => `${_getGraphemes(value)}`
+        ['123456', '12345 ', '123456789', 'foo bar baz'],
+        (value) => `${_getGraphemeCount(value)}`
+      );
+    });
+
+    test('for invalid emoji', () => {
+      expectActionIssue(
+        action,
+        baseIssue,
+        ['😀👋🏼🧩👩🏻‍🏫🫥🫠', '😀👋🏼🧩👩🏻‍🏫🫥🫠🧑‍💻👻🥎'],
+        (value) => `${_getGraphemeCount(value)}`
       );
     });
   });