+++ library/src/actions/notGraphemes/notGraphemes.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, test } from 'vitest';
-import { _getGraphemes } from '../../utils/index.ts';
+import type { StringIssue } from '../../schemas/index.ts';
+import { _getGraphemeCount } from '../../utils/index.ts';
 import { expectActionIssue, expectNoActionIssue } from '../../vitest/index.ts';
 import {
   notGraphemes,
@@ -16,7 +17,7 @@ describe('notGraphemes', () => {
       expects: '!5',
       requirement: 5,
       async: false,
-      _run: expect.any(Function),
+      '~validate': expect.any(Function),
     };
 
     test('with undefined message', () => {
@@ -48,19 +49,44 @@ describe('notGraphemes', () => {
     const action = notGraphemes(5);
 
     test('for untyped inputs', () => {
-      expect(action._run({ typed: false, value: null }, {})).toStrictEqual({
+      const issues: [StringIssue] = [
+        {
+          kind: 'schema',
+          type: 'string',
+          input: null,
+          expected: 'string',
+          received: 'null',
+          message: 'message',
+        },
+      ];
+      expect(
+        action['~validate']({ typed: false, value: null, issues }, {})
+      ).toStrictEqual({
         typed: false,
         value: null,
+        issues,
       });
     });
 
     test('for valid strings', () => {
       expectNoActionIssue(action, [
-        '😀😀😀😀',
+        '',
+        ' ',
+        '1',
         '1234',
-        'foobar',
-        '😀😀😀😀😀😀',
-        '🧑‍💻',
+        '123 ',
+        '123456',
+        '12 456',
+        '123456789',
+      ]);
+    });
+
+    test('for valid emoji', () => {
+      expectNoActionIssue(action, [
+        '😀',
+        '😀👋🏼🧩👩🏻‍🏫',
+        '😀👋🏼🧩👩🏻‍🏫🫥🫠',
+        '😀👋🏼🧩👩🏻‍🏫🫥🫠🧑‍💻👻🥎',
       ]);
     });
   });
@@ -82,8 +108,17 @@ describe('notGraphemes', () => {
       expectActionIssue(
         action,
         baseIssue,
-        ['😀😀😀😀😀', '12345', 'hello'],
-        (value) => `${_getGraphemes(value)}`
+        ['12345', '12 45', '1234 ', 'hello'],
+        (value) => `${_getGraphemeCount(value)}`
+      );
+    });
+
+    test('for invalid emoji', () => {
+      expectActionIssue(
+        action,
+        baseIssue,
+        ['😀👋🏼🧩👩🏻‍🏫🫥'],
+        (value) => `${_getGraphemeCount(value)}`
       );
     });
   });