+++ library/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## v1.0.0 (Month DD, YYYY)
 
 - Add `checkItemsAsync` action (pull request #856)
+- Add `graphemes`, `maxGraphemes`, `minGraphemes` and `notGraphemes` action (pull request #853)
 - Change types and implementation to support Standard Schema
 
 ## v0.42.1 (September 20, 2024)